refactor: moved render method to helper class
