+++ phpmyfaq/src/Bootstrap.php
@@ -26,7 +26,7 @@ use phpMyFAQ\Init;
 // Debug mode:
 // - false debug mode disabled
 // - true  debug mode enabled
-define('DEBUG', false);
+define('DEBUG', true);
 if (DEBUG) {
     ini_set('display_errors', 1);
     ini_set('display_startup_errors', 1);
